Rename task_queue.py to task_submit.py.
