+++ lms/djangoapps/instructor/views.py
@@ -25,7 +25,7 @@ from xmodule.modulestore.django import modulestore
 from xmodule.modulestore.exceptions import ItemNotFoundError
 
 from courseware import grades
-from courseware import task_queue
+from courseware import task_submit
 from courseware.access import (has_access, get_access_group_name,
                                course_beta_test_group_name)
 from courseware.courses import get_course_with_access
@@ -243,7 +243,7 @@ def instructor_dashboard(request, course_id):
         problem_urlname = request.POST.get('problem_for_all_students', '')
         problem_url = get_module_url(problem_urlname)
         try:
-            course_task_log_entry = task_queue.submit_rescore_problem_for_all_students(request, course_id, problem_url)
+            course_task_log_entry = task_submit.submit_rescore_problem_for_all_students(request, course_id, problem_url)
             if course_task_log_entry is None:
                 msg += '<font color="red">Failed to create a background task for rescoring "{0}".</font>'.format(problem_url)
             else:
@@ -260,7 +260,7 @@ def instructor_dashboard(request, course_id):
         problem_urlname = request.POST.get('problem_for_all_students', '')
         problem_url = get_module_url(problem_urlname)
         try:
-            course_task_log_entry = task_queue.submit_reset_problem_attempts_for_all_students(request, course_id, problem_url)
+            course_task_log_entry = task_submit.submit_reset_problem_attempts_for_all_students(request, course_id, problem_url)
             if course_task_log_entry is None:
                 msg += '<font color="red">Failed to create a background task for resetting "{0}".</font>'.format(problem_url)
             else:
@@ -356,7 +356,7 @@ def instructor_dashboard(request, course_id):
                     msg += "<font color='red'>Couldn't reset module state.  </font>"
             else:
                 try:
-                    course_task_log_entry = task_queue.submit_rescore_problem_for_student(request, course_id, module_state_key, student)
+                    course_task_log_entry = task_submit.submit_rescore_problem_for_student(request, course_id, module_state_key, student)
                     if course_task_log_entry is None:
                         msg += '<font color="red">Failed to create a background task for rescoring "{0}" for student {1}.</font>'.format(module_state_key, unique_student_identifier)
                     else:
@@ -721,7 +721,7 @@ def instructor_dashboard(request, course_id):
 
     # generate list of pending background tasks
     if settings.MITX_FEATURES.get('ENABLE_COURSE_BACKGROUND_TASKS'):
-        course_tasks = task_queue.get_running_course_tasks(course_id)
+        course_tasks = task_submit.get_running_course_tasks(course_id)
     else:
         course_tasks = None
 
@@ -1294,7 +1294,7 @@ def get_background_task_table(course_id, problem_url, student=None):
     Returns a tuple of (msg, datatable), where the msg is a possible error message,
     and the datatable is the datatable to be used for display.
     """
-    history_entries = task_queue.get_course_task_history(course_id, problem_url, student)
+    history_entries = task_submit.get_course_task_history(course_id, problem_url, student)
     datatable = None
     msg = ""
     # first check to see if there is any history at all
@@ -1329,7 +1329,7 @@ def get_background_task_table(course_id, problem_url, student=None):
                 if 'duration_ms' in task_outputs:
                     duration_ms = task_outputs['duration_ms']
             # get progress status message:
-            success, message = task_queue.get_task_completion_message(course_task)
+            success, message = task_submit.get_task_completion_message(course_task)
             if success:
                 status = "Complete"
             else:
