+++ lms/envs/test.py
@@ -189,9 +189,3 @@ PASSWORD_HASHERS = (
     # 'django.contrib.auth.hashers.CryptPasswordHasher',
 )
 
-################################# CELERY ######################################
-
-# By default don't use a worker, execute tasks as if they were local functions
-CELERY_ALWAYS_EAGER = True
-CELERY_RESULT_BACKEND = 'cache'
-BROKER_TRANSPORT = 'memory'
